Implement shuffle in queue.c
shuffle has implemented in queue.c, and add do_shuffle()
and add command in qtest.c, so that users are able to
test the shuffle function in qtest command line. However,
editing queue.h is not allowed. To pass the compile,
some of code are turn into the comments temporarily.

Change-Id: Iac458f53f9c9f2672fc5ee30e4854151cae99fd0